Betting Operations

+++ Sources/BetSwift/Operations/PlaceExecutionReport.swift
@@ -42,8 +42,8 @@ public struct PlaceExecutionReport: Codable, EndpointType {
     public let instruction: PlaceInstruction
     public let betId: String?
     public let placedDate: Date?
-    public let averagePriceMatched: Float?
-    public let sizeMatched: Float?
+    public let averagePriceMatched: Double?
+    public let sizeMatched: Double?
   }
   public let customerRef: String?
   public let status: ExecutionReportStatus
